Thursday Changes - Ring/Global AWP Auto Done
-- autonnumber now has a boolean variable with it called globalAuton that decides whether or not the auto route that is run should be the universal one for that quadrant or not
-- PIDMover now uses doubles for custom code during PID movements
-- different AWP auto code made for each quadrant (each in their own function)

+++ src/pid.cpp
@@ -5,7 +5,7 @@ void PIDMover(
 		bool reverse, // defaults to false- explicitly set to true to reverse the robot
 
 		std::vector<std::function<void(void)>> customs, // a lambda function that will execute during the PID (optional)
-		std::vector<int> executeAts // the distance point (in inches) that you want to trigger the custom lambda function at (optional)
+		std::vector<double> executeAts // the distance point (in inches) that you want to trigger the custom lambda function at (optional)
 		)
 {
